+++ compiler/bsb/bsb_exception.ml
@@ -47,7 +47,7 @@ let print (fmt : Format.formatter) (x : error) =
       modname
   | Package_not_found name ->
     let name = Bsb_pkg_types.to_string name in
-    if Ext_string.equal name !Bs_version.package_name then
+    if Ext_string.equal name Bs_version.package_name then
       Format.fprintf fmt
         "File \"rescript.json\", line 1\n\
          @{<error>Error:@} package @{<error>%s@} is not found\n\