+++ compiler/bsb/bsb_pkg_types.ml
@@ -83,9 +83,6 @@ let string_as_package (s : string) : t =
   if v = '@' then (
     let scope_id = Ext_string.no_slash_idx s in
     assert (scope_id > 0);
-    (* better-eror message for invalid scope package:
-       @rescript/std
-    *)
     Scope
       (String.sub s (scope_id + 1) (len - scope_id - 1), String.sub s 0 scope_id))
   else Global s