+++ compiler/bsc/rescript_compiler_main.ml
@@ -50,19 +50,6 @@ let set_abs_input_name sourcefile =
   sourcefile
 let setup_outcome_printer () = Lazy.force Res_outcome_printer.setup
 
-let setup_runtime_path path =
-  let u0 = Filename.dirname path in
-  let std = Filename.basename path in
-  let _path = Filename.dirname u0 in
-  let rescript = Filename.basename u0 in
-  (match rescript.[0] with
-  | '@' ->
-    (* scoped package *)
-    Bs_version.package_name := rescript ^ "/" ^ std
-  | _ -> Bs_version.package_name := std
-  | exception _ -> Bs_version.package_name := std);
-  Js_config.customize_runtime := Some path
-
 let process_file sourcefile ?kind ppf =
   (* This is a better default then "", it will be changed later
      The {!Location.input_name} relies on that we write the binary ast
@@ -415,9 +402,6 @@ let command_line_flags : (string * Bsc_args.spec * string) array =
       "<list>  Enable or disable error status for warnings according\n\
        to <list>.  See option -w for the syntax of <list>.\n\
        Default setting is " ^ Bsc_warnings.defaults_warn_error );
-    ( "-runtime",
-      string_call setup_runtime_path,
-      "*internal* Set the runtime directory" );
     ( "-make-runtime",
       unit_call Js_packages_state.make_runtime,
       "*internal* make runtime library" );