+++ compiler/common/bs_version.mli
@@ -26,4 +26,4 @@ val version : string
 
 val header : string
 
-val package_name : string ref
+val package_name : string